refactor: factored out ParameterReadingButton
